+++ composeApp/src/desktopMain/kotlin/eu/heha/samayouwa/SamayouwaDesktop.kt
@@ -1,8 +1,8 @@
 package eu.heha.samayouwa
 
-import eu.heha.samayouwa.model.DataStoreSettingsDao
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import eu.heha.samayouwa.model.DataStoreSettingsDao
 import kotlinx.io.files.Path
 import org.jetbrains.compose.resources.painterResource
 import samayouwa.composeapp.generated.resources.Res
@@ -13,10 +13,17 @@ fun main() {
     App.initialize(
         App.Requirements(
             settingsDaoFactory = {
-                val path = File("data")
+                val jarFilePath = App::class.java.protectionDomain.codeSource.location.file
+                    .replace("%20", " ")// as an uri it escapes spaces TODO: automate decode url
+
+                val rootFile = File(jarFilePath)
+                    .parentFile //app folder
+                    .parentFile //root folder
+
+                val dataFolder = File(rootFile, "data")
                     .also { it.mkdirs() }
-                    .let { Path(it.path) }
-                DataStoreSettingsDao(path)
+
+                DataStoreSettingsDao(Path(dataFolder.path))
             }
         )
     )