+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.1] - 2021-12-10
+
+- fixed settings not being saved on macos
+
 ## [1.1.0] - 2021-12-09
 
 - pressing space now toggles play/pause
@@ -12,12 +16,13 @@
 - added desktop icons
 - added persistence of settings
 - overhauled build process for desktop apps
+- tested for windows and macos
 
 ## [1.0.0] - 2024-11-27
 
 - basic interpolation of a ball going from right to left and back in a linear fashion
 - added play/pause functionality
-- added set~~~~tings for
+- added settings for
     - velocity
     - size of the ball
     - color of the ball